// src/apis/models/Volo_Abp_TextTemplateManagement_TextTemplates_UpdateTemplateContentInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_Abp_TextTemplateManagement_TextTemplates_UpdateTemplateContentInput = {
    templateName: string;
    cultureName?: string | null;
    content?: string | null;
};